hash list completely

+++ MyCode/list.cpp
@@ -11,41 +11,15 @@ List::List(const List &list){
     head.prev = list.head.prev;
 }
 
-List::List(){
-    compare_func = default_list_compare;
-    destroy_func = default_list_destroy;
-    init_list_head();
-}
 
 List::List(pfn_list_compare compare,pfn_list_destroy destroy){
-    if(compare != NULL){
-        compare_func = compare;
-    }else{
-        compare_func = default_list_compare;
-    }
-    if(destroy != NULL){
-        destroy_func = destroy;
-    }else{
-        destroy_func = default_list_destroy;
-    }
+    compare_func = compare;
+    destroy_func = destroy;
     init_list_head();
 }
 
 List::~List(){
     release_list();
-}
-
-void List::set_fpn(pfn_list_compare compare,pfn_list_destroy destroy){
-    if(compare != NULL){
-        compare_func = compare;
-    }else{
-        compare_func = default_list_compare;
-    }
-    if(destroy != NULL){
-        destroy_func = destroy;
-    }else{
-        destroy_func = default_list_destroy;
-    }
 }
 
 void List::init_list_item(list_head_t *item){
@@ -71,7 +45,9 @@ void List::release_list(){
     list_head_t *tmp_head = head.next; 
     while(tmp_head != &head){ 
         remove_list_item_(tmp_head->prev,tmp_head->next,tmp_head);
-        destroy_func(tmp_head->data);
+        if(destroy_func != NULL){
+            destroy_func(tmp_head->data);
+        }
         free(tmp_head);
         tmp_head = head.next;
     } 
@@ -89,8 +65,13 @@ void List::dump_list(){
 void* List::find_list_item(void *data){
     list_head_t *tmp = head.next ; 
     while(tmp != &head){
-        if(compare_func(tmp->data,data) == 0){
-            return tmp->data;
+        if(compare_func != NULL){
+            if(compare_func(tmp->data,data) == 0){
+                return tmp->data;
+            }
+        }else{
+            printf("no compare function for the list \n");
+            break;
         }
         tmp  = tmp->next;
     }
@@ -128,14 +109,19 @@ void List::insert_list_head(void *data){
 void* List::get_prev_item(void *data){
     list_head_t *tmp = head.next ; 
     while(tmp != &head){
-        if(compare_func(tmp->data,data) == 0){
-            if(tmp->prev != &head){
-                return tmp->prev->data;
-            }else{
-                return NULL;
+        if(compare_func != NULL){
+            if(compare_func(tmp->data,data) == 0){
+                if(tmp->prev != &head){
+                    return tmp->prev->data;
+                }else{
+                    return NULL;
+                }
             }
+            tmp  = tmp->next;
+        }else{
+            printf("no compare function provided \n");
+            return NULL;
         }
-        tmp  = tmp->next;
     }
     if(&head != head.prev){
         return head.prev->data;
@@ -145,14 +131,19 @@ void* List::get_prev_item(void *data){
 void* List::get_next_item(void *data){
     list_head_t *tmp = head.next ; 
     while(tmp != &head){
-        if(compare_func(tmp->data,data) == 0){
-            if(tmp->next != &head){
-                return tmp->next->data;
-            }else{
-                return NULL;
+        if(compare_func != NULL){
+            if(compare_func(tmp->data,data) == 0){
+                if(tmp->next != &head){
+                    return tmp->next->data;
+                }else{
+                    return NULL;
+                }
             }
+            tmp  = tmp->next;
+        }else{
+            printf("no compare function provided \n");
+            return NULL;
         }
-        tmp  = tmp->next;
     }
     if(&head != head.next){
         return head.next->data;
@@ -163,14 +154,19 @@ void* List::get_next_item(void *data){
 void* List::insert_list_local(void *prev_data,void *data){
     list_head_t *tmp = head.next ; 
     while(tmp != &head){
-        if(compare_func(tmp->data,data) == 0){
-            list_head_t *new_item = get_new_item();
-            assert(new_item != NULL);
-            new_item->data = data;
-            add_list_item_(tmp,tmp->next,new_item);
-            return data;
+        if(compare_func != NULL){
+            if(compare_func(tmp->data,data) == 0){
+                list_head_t *new_item = get_new_item();
+                assert(new_item != NULL);
+                new_item->data = data;
+                add_list_item_(tmp,tmp->next,new_item);
+                return data;
+            }
+            tmp  = tmp->next;
+        }else{
+            printf("no compare function provided \n");
+            return NULL;
         }
-        tmp  = tmp->next;
     }
     return NULL;
 }
@@ -178,13 +174,20 @@ void* List::insert_list_local(void *prev_data,void *data){
 void List::remove_list_item(void *data){
     list_head_t *tmp = head.next ; 
     while(tmp != &head){
-        if(compare_func(tmp->data,data) == 0){
-            remove_list_item_(tmp->prev,tmp->next,tmp);
-            destroy_func(tmp->data);
-            free(tmp);
-            return;
+        if(compare_func != NULL){
+            if(compare_func(tmp->data,data) == 0){
+                remove_list_item_(tmp->prev,tmp->next,tmp);
+                if(destroy_func != NULL){
+                    destroy_func(tmp->data);
+                }
+                free(tmp);
+                return;
+            }
+            tmp  = tmp->next;
+        }else{
+            printf("no compare function provided \n");
+            return ;
         }
-        tmp  = tmp->next;
     }
 }
